Oculus and Kinect
git-svn-id: https://mediacenter-pc/svn/GPGPU/CUDA/RaytracingEngine/trunk@853 c4de0f11-4f5c-764a-a4aa-d9e147179871

+++ OBJReader.cpp
@@ -1,3 +1,25 @@
+/* 
+* Copyright (C) 2011-2012 Cyrille Favreau <cyrille_favreau@hotmail.com>
+*
+* This library is free software; you can redistribute it and/or
+* modify it under the terms of the GNU Library General Public
+* License as published by the Free Software Foundation; either
+* version 2 of the License, or (at your option) any later version.
+*
+* This library is distributed in the hope that it will be useful,
+* but WITHOUT ANY WARRANTY; without even the implied warranty of
+* MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the GNU
+* Library General Public License for more details.
+*
+* You should have received a copy of the GNU General Public License
+* along with this program.  If not, see <http://www.gnu.org/licenses/>. 
+*/
+
+/*
+* Author: Cyrille Favreau <cyrille_favreau@hotmail.com>
+*
+*/
+
 #include <fstream>
 #include <map>
 #include <vector>
@@ -192,7 +214,7 @@ unsigned int OBJReader::loadMaterialsFromFile(
          {
             // Specular values
             line = line.substr(2);
-            float d=atof(line.c_str());
+            float d=static_cast<float>(atof(line.c_str()));
             m_materials[id].reflection   = 1.f; 
             m_materials[id].transparency = d; 
             m_materials[id].refraction   = 1.66;
